auto_spi_soc_prog_index_git.py Creation
The aim of this file is to load fast all data available, skipping the series where is no data. It is very likely that in the future when revising the data, revising of the code below is also needed, because the link to the file was gotten after request, and also it has some modifications from year to year.